refactor #148 : 사용하지 않는 폴더 및 파일 제거(DataTestViewModel가 제거되어서 사용하던 곳에서 추가 viewModel 구현 필요)
