+++ Outline/Outline.xcodeproj/project.xcworkspace/xcshareddata/swiftpm/Package.resolved
@@ -12,7 +12,7 @@
     {
       "identity" : "firebase-ios-sdk",
       "kind" : "remoteSourceControl",
-      "location" : "https://github.com/firebase/firebase-ios-sdk.git",
+      "location" : "https://github.com/firebase/firebase-ios-sdk",
       "state" : {
         "revision" : "837d4af6ead57cec1fc38007892500d3139c7556",
         "version" : "10.16.0"
