--- Outline/Outline/Source/DesignSystem/Component/Component.swift
@@ -1,8 +1,0 @@
-//
-//  Component.swift
-//  Outline
-//
-//  Created by Hyunjun Kim on 10/15/23.
-//
-
-import Foundation
