--- Outline/Outline/Source/DesignSystem/Modifier/Modifier.swift
@@ -1,8 +1,0 @@
-//
-//  Modifier.swift
-//  Outline
-//
-//  Created by Hyunjun Kim on 10/13/23.
-//
-
-import Foundation
